cancel button download

+++ admin/include/dp-wallpaper-downloader.php
@@ -2,21 +2,25 @@
 
 if (isset($_POST['dp-downloader-status'])) {
     if ($_POST['dp-download-status'] == 1) {
+        $waktu = date("dmy-His"); // create folder
         update_option('dp-download-status', '1');
         update_option('dp-download-keywords', $_POST['dp-keywords']);
         update_option('dp-download-total', $_POST['dp-download-perkeyword']);
+        update_option('dp-download-dir', "$waktu");
         echo '<meta http-equiv="refresh" content="0">'; // refresh browser
         if (! wp_next_scheduled ( 'dp_download_keywords' )) { // do the download process using wp cron
-            wp_schedule_single_event( time(), 'dp_download_keywords' );
+            wp_schedule_single_event( time() + 10, 'dp_download_keywords' );
+        } else {
+            wp_clear_scheduled_hook( 'dp_download_keywords' );
         }
     } else if ($_POST['dp-download-status'] == 0) {
         update_option('dp-download-status', '0');
-        update_option('dp-download-keywords', '');
-        update_option('dp-download-total', '');
         echo '<meta http-equiv="refresh" content="0">'; // refresh browser
         // remove wallpapers.txt & current canceled wallpaper directory
         $uploadDir = wp_upload_dir()['basedir'];
-        @unlink("$uploadDir/wallpapers.txt");
+        if (file_exists("$uploadDir/wallpapers.txt")) {
+            unlink("$uploadDir/wallpapers.txt");
+        }
         $wallDir = get_option('dp-download-dir');
         if (file_exists("$uploadDir/$wallDir/")) {
             array_map('unlink', glob("$uploadDir/$wallDir/*.*"));
